LayoutTests/imported/w3c/web-platform-tests/css/css-scroll-snap/scroll-initial-target/scroll-initial-target-display-toggled.tentative-expected.txt


FAIL display:block scroll-initial-target becomes display: none assert_equals: inner-scroller is scrolled to scroll-initial-target expected 317 but got 0
FAIL display:none scroll-initial-target becomes display: block assert_equals: inner-scroller is scrolled to scroll-initial-target expected 317 but got 0
